Fixed an additional issue
* Fixed the Frame Rate setting for mode 1.
* Fixed a syntax error left in main.cpp

+++ src/Frame.h
@@ -59,7 +59,7 @@ class Frame {
         int                 mFrameNumber;        // Each frame is identified by a number corresponding to the acquisition order.
         int                 mFrameRemaining;     // Define the number of remaining frames if the input source is a video or a set of single frames.
         double              mSaturatedValue;     // Max pixel value in the image.
-        int                 mFps;                // Camera's fps.
+        double              mFps;                // Camera's fps.
         int                 mWidth;
         int                 mHeight;
